+++ inference-engine/src/vpu/graph_transformer/src/middleend/passes/merge_permute_stages.cpp
@@ -1,4 +1,4 @@
-// Copyright (C) 2018-2020 Intel Corporation
+// Copyright (C) 2018-2021 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 //
 
@@ -58,7 +58,7 @@ private:
     static bool isTrivialPermute(const PermutationIndexVector& permutation, const vpu::DimValues& dims) {
         InferenceEngine::SizeVector dimsVector(dims.size());
         for (const auto& dim : dims) {
-            auto index = dimToIeInd(dim.first, dims.size());
+            auto index = dimToIeInd(dim.first, static_cast<int>(dims.size()));
             dimsVector[dims.size() - 1 - index] = dim.second;
         }
